// sample-ui/src/api/interfaces/AppRoleAssignment.ts
/* tslint:disable */

/**
 * This file was automatically generated by "Swaxios".
 * It should not be modified by hand.
 */

export interface AppRoleAssignment {
  appRoleId?: string;
  createdDateTime?: string;
  principalDisplayName?: string;
  principalId?: string;
  principalType?: string;
  resourceDisplayName?: string;
  resourceId?: string;
  deletedDateTime?: string;
  id?: string;
  _odata_type?: string;
}